+++ auto_save_restrict/static/src/js/form_controller.js
@@ -1,33 +1,54 @@
-/** @odoo-module */
+/** @odoo-module **/
+
 import { FormController } from "@web/views/form/form_controller";
 import { patch } from "@web/core/utils/patch";
-import { useSetupView } from "@web/views/view_hook";
+import { _t } from "@web/core/l10n/translation";
+import { SettingsConfirmationDialog } from "@web/webclient/settings_form_view/settings_confirmation_dialog";
+
+// Save original method reference
+const _superOnPagerUpdate = FormController.prototype.onPagerUpdate;
+
 patch(FormController.prototype, {
-/* Patch FormController to restrict auto save in form views */
-   setup(){
-      super.setup(...arguments);
-      this.beforeLeaveHook = false
-      useSetupView({
-          beforeLeave: () => this.beforeLeave(),
-          beforeUnload: (ev) => this.beforeUnload(ev),
-      });
-   },
-   async beforeLeave() {
-   /* function will work before leave the form */
-      if(this.model.root.isDirty && this.beforeLeaveHook == false){
-          if (confirm("Do you want to save changes before leaving?")) {
-              this.beforeLeaveHook = true
-              await this.model.root.save({
-                  reload: false,
-                  onError: this.onSaveError.bind(this),
-              });
-          } else {
-              this.beforeLeaveHook = true
-              this.model.root.discard();
-          }
-      }
-   },
-   beforeUnload: async (ev) => {
-       ev.preventDefault();
-   }
+    async beforeLeave() {
+        const dirty = await this.model.root.isDirty();
+        if (dirty) {
+            return this._confirmSave();
+        }
+        return true;
+    },
+
+    beforeUnload() {},
+
+    async _confirmSave() {
+        let _continue = true;
+        await new Promise((resolve) => {
+            this.dialogService.add(SettingsConfirmationDialog, {
+                body: _t("Would you like to save your changes?"),
+                confirm: async () => {
+                    await this.save();
+                    _continue = true;
+                    resolve();
+                },
+                cancel: async () => {
+                    await this.model.root.discard();
+                    _continue = true;
+                    resolve();
+                },
+                stayHere: () => {
+                    _continue = false;
+                    resolve();
+                },
+            });
+        });
+        return _continue;
+    },
+
+    async onPagerUpdate(value) {
+        const proceed = await this.beforeLeave();
+        if (!proceed) {
+            return;
+        }
+        await this.model.root.discard();
+        return _superOnPagerUpdate.call(this, value);
+    },
 });