+++ custom_receipts_for_pos/static/src/js/receipt_design.js
@@ -13,12 +13,17 @@ patch(OrderReceipt.prototype, {
         this.pos = useState(useService("pos"));
     },
     get templateProps() {
+        const order = this.pos.get_order();
+        const partner = order ? order.get_partner() : null;
+
         return {
+            pos:this.pos,
             data: this.props.data,
-            order: this.pos.orders,
-            receipt: this.pos.orders[0].export_for_printing(),
-            orderlines:this.pos.orders[0].get_orderlines(),
-            paymentlines:this.pos.orders[0].get_paymentlines()
+            order: order,
+            receipt:this.props.data,
+            orderlines:this.props.data.orderlines,
+            paymentlines:this.props.data.paymentlines,
+            partner: partner,
         };
     },
     get templateComponent() {