+++ pos_access_right_hr/static/src/js/ActionpadWidget.js
@@ -15,7 +15,6 @@ patch(ActionpadWidget.prototype, {
         }
     },
     disable_payment() {
-        console.log(this.pos.cashier?.disable_customer)
         if (this.pos.cashier?.disable_payment) {
             return true;
         } else {