+++ pos_restrict_product_stock/static/src/js/RestrictStockPopup.js
@@ -5,12 +5,17 @@
 import { AbstractAwaitablePopup } from "@point_of_sale/app/popup/abstract_awaitable_popup";
 
 class RestrictStockPopup extends AbstractAwaitablePopup {
+
     _OrderProduct() {
     // On clicking order product button on popup, it will add product to orderline
-        var product = this.env.services.pos.db.get_product_by_id(this.props.pro_id)
-        this.env.services.pos.selectedOrder.add_product(product);
-        this.cancel();
-    }
+        if(this.props.pro_id){
+            var product = this.env.services.pos.db.get_product_by_id(this.props.pro_id)
+            product.order_status = true
+            this.env.services.pos.selectedOrder.add_product(product);
+        }
+            this.props.resolve(true);
+            this.props.close();
+        }
 }
 RestrictStockPopup.template = 'RestrictStockPopup';
 export default RestrictStockPopup;