+++ product_multi_uom_pos/static/src/overrides/models/orderline.js
@@ -51,15 +51,20 @@ patch(Orderline.prototype, {
             const currentOrder = this.env.services.pos.get_order();
             currentOrder.selected_orderline.set_uom({0:uomId,1:uomName})
           // Set the price_manually_set flag to indicate that the price was manually set
-            currentOrder.selected_orderline.price_type = "manual";
+            currentOrder.selected_orderline.price_manually_set = true;
            // Set the unit price of selected UoM on the order line
             currentOrder.selected_orderline.set_unit_price(price);
     },
      getUom(self) {
         const currentOrder = self.env.services.pos.get_order();
         const currentLine = currentOrder.orderlines.find((line) => line.full_product_name === self.props.line.productName)
-        const uom = currentLine.product.pos_multi_uom_ids
-        const filteredData = self.env.services.pos.pos_multi_uom.filter(obj => currentLine.product.pos_multi_uom_ids.includes(obj.id));
+        var filteredData = {}
+        if(currentLine){
+            const uom = currentLine.product.pos_multi_uom_ids
+            if(uom){
+            filteredData = self.env.services.pos.pos_multi_uom.filter(obj => currentLine.product.pos_multi_uom_ids.includes(obj.id));
+            }
+        }
         return filteredData;
     },
     resetUom(){