+++ auto_save_restrict/doc/RELEASE_NOTES.md
@@ -15,3 +15,13 @@
 #### Version 17.0.1.0.2
 ##### BUG FIX
 - Fixed the issue in the form where the popup appeared repeatedly
+
+#### 31.05.2025
+#### Version 17.0.1.0.3
+##### BUG FIX
+- Fixed the issue in the smart button where the changes save automatically.
+
+#### 02.07.2025
+#### Version 17.0.1.0.4
+##### BUG FIX
+- Fixed the issue in the next/previous button where the changes save automatically.