+++ invoice_stock_move/doc/RELEASE_NOTES.md
@@ -5,3 +5,13 @@
 #### ADD
 
 Initial Commit for Stock Picking From Invoice
+
+#### 12.09.2024
+#### Version 17.0.1.0.1
+#### Bugfix
+-An issue was mentioned that picking type value is missing , updated the code
+
+#### 06.03.2025
+#### Version 17.0.1.0.2
+#### Bugfix
+-An issue was mentioned that picking type selection and user access , updated the code