+++ onedrive_integration_odoo/doc/RELEASE_NOTES.md
@@ -3,3 +3,13 @@
 #### Version 17.0.1.0.0
 #### ADD
 - Initial commit for Onedrive Integration
+
+#### 16.08.2024
+#### Version 17.0.1.0.1
+#### UPDT
+-  Added modules field and changed view
+
+#### 29.03.2025
+#### Version 17.0.1.0.2
+#### BUG FIX
+- Added /consumers endpoint