+++ sale_product_image/doc/RELEASE_NOTES.md
@@ -5,3 +5,15 @@
 ##### ADD
 
 - Initial commit for Sale Order Line Images
+
+#### 10.12.2024
+#### Version 17.0.1.0.0
+##### Bug fixes
+
+- Updated the image size in the report
+
+#### 10.12.2024
+#### Version 17.0.1.1.0
+##### Bug fixes
+
+- Updated the image in the portal sale orders